traffic-agent/requirements.txt
"""
Archivo de requisitos para el backend Flask.
"""
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.267
langchain-openai==0.0.2
# Eliminado langchain-core==0.1.1 para permitir que langchain-openai gestione la versión compatible
openai==1.3.5
rapidfuzz==3.5.2
nltk==3.8.1
